# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档 for Aurora 项目",
        "title": "Aurora API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/api/current_user": {
            "get": {
                "description": "获取当前登录用户的信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前用户信息",
                "responses": {
                    "200": {
                        "description": "获取用户信息成功",
                        "schema": {
                            "$ref": "#/definitions/routes.currentUserResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取用户信息失败",
                        "schema": {
                            "$ref": "#/definitions/routes.currentUserResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/has_signed": {
            "get": {
                "description": "获取当前用户今日的签到状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取签到状态",
                "responses": {
                    "200": {
                        "description": "获取签到状态成功",
                        "schema": {
                            "$ref": "#/definitions/routes.hasSignedResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "获取签到状态失败",
                        "schema": {
                            "$ref": "#/definitions/routes.hasSignedResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "通过邮箱和密码进行用户登录验证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/routes.loginResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/routes.loginResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "description": "清除用户登录会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "退出成功",
                        "schema": {
                            "$ref": "#/definitions/routes.logoutResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "退出失败",
                        "schema": {
                            "$ref": "#/definitions/routes.logoutResponseFailed"
                        }
                    }
                }
            }
        },
...
        "/api/send_verify_code": {
            "post": {
                "description": "向指定邮箱发送验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "邮箱地址",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.sendVerifyCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码发送成功",
                        "schema": {
                            "$ref": "#/definitions/routes.sendVerifyCodeResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "验证码发送失败",
                        "schema": {
                            "$ref": "#/definitions/routes.sendVerifyCodeResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/sign": {
            "post": {
                "description": "用户每日签到获取积分",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户签到",
                "responses": {
                    "200": {
                        "description": "签到成功",
                        "schema": {
                            "$ref": "#/definitions/routes.signResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "签到失败",
                        "schema": {
                            "$ref": "#/definitions/routes.signResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "description": "通过用户名、邮箱、密码和验证码进行用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "signup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.signupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/routes.signupResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/routes.signupResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/verify_points": {
            "post": {
                "description": "通过订单ID验证并充值积分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "验证积分充值",
                "parameters": [
                    {
                        "description": "验证信息",
                        "name": "verifyPoints",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.verifyPointsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "积分验证成功",
                        "schema": {
                            "$ref": "#/definitions/routes.verifyPointsResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "积分验证失败",
                        "schema": {
                            "$ref": "#/definitions/routes.verifyPointsResponseFailed"
                        }
                    }
                }
            }
        },
        "/api/verify_vip": {
            "post": {
                "description": "通过订单ID验证并激活VIP会员资格",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "验证VIP会员",
                "parameters": [
                    {
                        "description": "验证信息",
                        "name": "verifyVip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.verifyVipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "VIP验证成功",
                        "schema": {
                            "$ref": "#/definitions/routes.verifyVipResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "VIP验证失败",
                        "schema": {
                            "$ref": "#/definitions/routes.verifyVipResponseFailed"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.currentUserResponseFailed": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.currentUserResponseSuccess": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "isMember": {
                    "type": "boolean",
                    "example": true
                },
                "memberLevel": {
                    "type": "string",
                    "example": "VIP"
                },
                "points": {
                    "type": "integer",
                    "example": 100
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "username": {
                    "type": "string",
                    "example": "currentuser"
                }
            }
        },
...
        "routes.hasSignedResponseFailed": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.hasSignedResponseSuccess": {
            "type": "object",
            "properties": {
                "signed": {
                    "type": "boolean",
                    "example": false
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "routes.loginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "routes.loginResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "用户名或密码错误"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.loginResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "登录成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "routes.logoutResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "退出失败"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.logoutResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "退出成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "routes.sendVerifyCodeRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "routes.sendVerifyCodeResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "发送失败"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.sendVerifyCodeResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "发送成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "routes.signResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "今日已签到"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.signResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "签到成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "routes.signupRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "newuser@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "newuser"
                },
                "verifyCode": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "routes.signupResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "该邮箱已被注册"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.signupResponseSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "注册成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/routes.signupUser"
                }
            }
        },
        "routes.signupUser": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "newuser@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "1234567890"
                },
                "username": {
                    "type": "string",
                    "example": "newuser"
                }
            }
        },
...
        "routes.verifyPointsRequest": {
            "type": "object",
            "properties": {
                "orderID": {
                    "type": "string",
                    "example": "订单ID"
                }
            }
        },
        "routes.verifyPointsResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "无效订单号"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.verifyPointsResponseSuccess": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "routes.verifyVipRequest": {
            "type": "object",
            "properties": {
                "force": {
                    "type": "boolean",
                    "example": false
                },
                "orderID": {
                    "type": "string",
                    "example": "订单ID"
                }
            }
        },
        "routes.verifyVipResponseFailed": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "无效订单号"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "routes.verifyVipResponseSuccess": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}